refactor(buildings): share façade bay packing and connectable sync
Extract `BaySpec` / `PlacedBay` packing and window/door catalogs into
`openings::facade_bays`, and add `sync_connectable_openings_from_mapped`
so Les Halles and I-Apartment no longer duplicate wall-run fit or
post-map Passage/Aperture sync.

## Agent Dialogue

- **Developer:** Do hygiene items 1 (shared façade bays) and 2 (shared connectable-openings sync helper).
- **Agent:** Added `openings/facade_bays.rs` + re-exports; rewired Les Halles aliases and I-Apartment to shared APIs; both floor plans call `sync_connectable_openings_from_mapped`. `cargo test -p richmond-buildings --lib` green (278).
